buildfastwithai/pyproject.toml
[project]
name = "build-fast-with-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "educhain>=0.3.10",
    "langchain-google-genai>=2.1.6",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
]
